allow custom method, headers through cli flags
The implementation for the ArbitraryHeader type started off as an
attempt at implementing the hyper::header::Header trait, but failed
because the header name returned by that trait cannot be based on the
values within the struct, since there is no reference to "self". Thus,
the implementation here simply holds the parsed header from the command
line for use in the request building process.

The implementation of header parsing was based on the
hyper::header::Host parsing logic:
https://docs.rs/hyper/*/src/hyper/header/common/host.rs.html#88-102